+++ babylab/__about__.py
@@ -2,4 +2,4 @@
 About the package.
 """
 
-__version__ = "0.2.0"  # no cov
+__version__ = "0.2.1"  # no cov