+++ babylab/src/api.py
@@ -4,10 +4,11 @@
 Functions to interact with the REDCap API.
 """
 
-import json
-import zipfile
+import math
 import os
 import re
+import json
+import zipfile
 import datetime
 from dataclasses import dataclass
 from collections import OrderedDict
@@ -47,6 +48,13 @@ class Participant:
             for k, v in data.items()
             if k.startswith("participant_") or k == "record_id"
         }
+        date_birth = get_birth_date(
+            str(data["age_created_months"]) + ":" + str(data["age_created_days"]),
+            datetime.datetime.strptime(data["date_created"], "%Y-%m-%d %H:%M:%S"),
+        )
+        data["age_now_months"], data["age_now_days"] = get_age(
+            date_birth, datetime.datetime.today()
+        )
         self.record_id = data["record_id"]
         self.data = data
         self.appointments = apt
@@ -60,7 +68,7 @@ class Participant:
         """
         n_apt = 0 if self.appointments is None else len(self.appointments.records)
         n_que = 0 if self.questionnaires is None else len(self.questionnaires.records)
-        return f"Participant {self.record_id}: {n_apt} appointments, {n_que} questionnaires"  # pylint: disable=line-too-long
+        return f"Participant {self.record_id}: {str(n_apt)} appointments, {str(n_que)} questionnaires"  # pylint: disable=line-too-long
 
     def __str__(self):
         """Return class description as string.
@@ -70,7 +78,7 @@ class Participant:
         """
         n_apt = 0 if self.appointments is None else len(self.appointments.records)
         n_que = 0 if self.questionnaires is None else len(self.questionnaires.records)
-        return f"Participant {self.record_id}: {n_apt} appointments, {n_que} questionnaires"  # pylint: disable=line-too-long
+        return f"Participant {self.record_id}: {str(n_apt)} appointments, {str(n_que)} questionnaires"  # pylint: disable=line-too-long
 
 
 class Appointment:
@@ -85,9 +93,7 @@ class Appointment:
         }
         self.record_id = data["record_id"]
         self.data = data
-        self.appointment_id = (
-            data["record_id"] + ":" + str(data["redcap_repeat_instance"])
-        )
+        self.appointment_id = make_id(data["record_id"], data["redcap_repeat_instance"])
         self.status = data["status"]
         self.date = data["date"]
 
@@ -118,9 +124,7 @@ class Questionnaire:
             if k.startswith("language_") or k in ["record_id", "redcap_repeat_instance"]
         }
         self.record_id = data["record_id"]
-        self.questionnaire_id = (
-            data["record_id"] + ":" + str(data["redcap_repeat_instance"])
-        )
+        self.questionnaire_id = make_id(self.record_id, data["redcap_repeat_instance"])
         self.isestimated = data["isestimated"]
         self.data = data
         for i in range(1, 5):
@@ -154,6 +158,34 @@ class Questionnaire:
             + f"\n- L3 ({self.data['lang3']}) = {self.data['lang3_exp']}%"
             + f"\n- L4 ({self.data['lang4']}) = {self.data['lang4_exp']}%"
         )  # pylint: disable=line-too-long
+
+
+class User:
+    """User class"""
+
+    def __init__(self, **kwargs):
+        fields = {"content": "user", "format": "json", "returnFormat": "json"}
+        r = post_request(fields, **kwargs)
+        user = json.loads(r.content.decode())[0]
+        self.user = user["username"]
+        self.name = user["firstname"] + " " + user["lastname"]
+        self.email = user["email"]
+
+    def __repr__(self):
+        """Print class in console.
+
+        Returns:
+            str: Description to print in console.
+        """
+        return f"User {self.user} ({self.name}, {self.email})"
+
+    def __str__(self):
+        """Return class description as string.
+
+        Returns:
+            str: Description of class.
+        """
+        return f"User {self.user} ({self.name}, {self.email})"
 
 
 class BadTokenException(Exception):
@@ -274,6 +306,18 @@ def datetimes_to_strings(data: dict):
     return data
 
 
+def get_next_id(**kwargs) -> str:
+    """Get next record_id in REDCap database.
+
+    Args:
+        **kwargs: Additional arguments passed to ``post_request``.
+    Returns:
+        str: record_id of next record.
+    """
+    fields = {"content": "generateNextRecordName"}
+    return str(post_request(fields=fields, **kwargs).json())
+
+
 def get_records(record_id: str | list = None, **kwargs):
     """Return records as JSON.
 
@@ -298,6 +342,36 @@ def get_records(record_id: str | list = None, **kwargs):
     records = post_request(fields=fields, **kwargs).json()
     records = [datetimes_to_strings(r) for r in records]
     return records
+
+
+def make_id(ppt_id: str, repeat_id: str = None):
+    """Make a record ID.
+
+    Args:
+        ppt_id (str): Participant ID.
+        repeat_id (str, optional): Appointment or Questionnaire ID, or ``redcap_repeated_id``. Defaults to None.
+
+    Returns:
+        str: Record ID.
+    """  # pylint: disable=line-too-long
+    ppt_id = str(ppt_id)
+    if not ppt_id.isdigit():
+        raise ValueError(f"`ppt_id`` must be a digit, but '{ppt_id}' was provided")
+    if repeat_id is None:
+        return ppt_id
+    repeat_id = str(repeat_id)
+    if not repeat_id.isdigit():
+        raise ValueError(
+            f"`repeat_id`` must be a digit, but '{repeat_id}' was provided"
+        )
+    return ppt_id + ":" + repeat_id
+
+
+class RecordNotFound(Exception):
+    """If record is not found."""
+
+    def __init__(self, record_id):
+        super().__init__(f"Record '{record_id}' not found")
 
 
 def get_participant(ppt_id: str, **kwargs):
@@ -336,8 +410,46 @@ def get_participant(ppt_id: str, **kwargs):
             apt[repeat_id] = Appointment(r)
         if form == "language":
             que[repeat_id] = Questionnaire(r)
-    ppt = Participant(recs[0], apt=RecordList(apt), que=RecordList(que))
-    return ppt
+    try:
+        return Participant(recs[0], apt=RecordList(apt), que=RecordList(que))
+    except IndexError as exc:
+        raise RecordNotFound(record_id=ppt_id) from exc
+
+
+def get_appointment(apt_id: str, **kwargs):
+    """Get appointment record.
+
+    Args:
+        apt_id: ID of appointment (``redcap_repeated_id``).
+        **kwargs: Additional arguments passed to ``post_request``
+
+    Returns:
+        api.Appointment: Appointment object.
+    """
+    ppt_id, _ = apt_id.split(":")
+    ppt = get_participant(ppt_id, **kwargs)
+    try:
+        return ppt.appointments.records[apt_id]
+    except KeyError as exc:
+        raise RecordNotFound(record_id=apt_id) from exc
+
+
+def get_questionnaire(que_id: str, **kwargs):
+    """Get questionnaire record.
+
+    Args:
+        que_id: ID of appointment (``redcap_repeated_id``).
+        **kwargs: Additional arguments passed to ``post_request``
+
+    Returns:
+        api.Questionnaire: Appointment object.
+    """
+    ppt_id, _ = que_id.split(":")
+    ppt = get_participant(ppt_id, **kwargs)
+    try:
+        return ppt.questionnaires.records[que_id]
+    except KeyError as exc:
+        raise RecordNotFound(record_id=que_id) from exc
 
 
 def add_participant(data: dict, modifying: bool = False, **kwargs):
@@ -524,18 +636,17 @@ class Records:
         appointments = {}
         questionnaires = {}
         for r in records:
-            if r["redcap_repeat_instance"] and r["appointment_status"]:
-                r["appointment_id"] = (
-                    r["record_id"] + ":" + str(r["redcap_repeat_instance"])
-                )
+            ppt_id = r["record_id"]
+            repeat_id = r["redcap_repeat_instance"]
+
+            if repeat_id and r["appointment_status"]:
+                r["appointment_id"] = make_id(ppt_id, repeat_id)
                 appointments[r["appointment_id"]] = Appointment(r)
-            if r["redcap_repeat_instance"] and r["language_lang1"]:
-                r["questionnaire_id"] = (
-                    r["record_id"] + ":" + str(r["redcap_repeat_instance"])
-                )
+            if repeat_id and r["language_lang1"]:
+                r["questionnaire_id"] = make_id(ppt_id, repeat_id)
                 questionnaires[r["questionnaire_id"]] = Questionnaire(r)
             if not r["redcap_repeat_instrument"]:
-                participants[r["record_id"]] = Participant(r)
+                participants[ppt_id] = Participant(r)
 
         # add appointments and questionnaires to each participant
         for p, v in participants.items():
@@ -662,5 +773,5 @@ def get_birth_date(age: str, timestamp: str | datetime.datetime = None):
     if not isinstance(timestamp, datetime.datetime):
         timestamp = datetime.datetime.strptime(timestamp, "%Y-%m-%d")
     age_parsed = age.split(":")
-    days_diff = int(float(age_parsed[0]) * 30.437 + float(age_parsed[1]))
+    days_diff = math.ceil(float(age_parsed[0]) * 30.437 + float(age_parsed[1]))
     return timestamp - relativedelta.relativedelta(days=days_diff)